app/pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "llm_feedback_classifier"
version = "0.0.1"
description = "a llm that classifies the client feedback"
authors = [
  { name = "Your name (or your organization/company/team)" },
]

readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    
]
dependencies = [
    "docker>=7.1.0",
    "fastapi>=0.136.0",
    "langchain-google-genai>=4.2.2",
    "langgraph>=1.1.8",
    "pydantic>=2.13.2",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "streamlit>=1.56.0",
    "tenacity>=9.1.4",
    "uvicorn>=0.44.0",
]
requires-python = "~=3.10.0"


[tool.ruff]
line-length = 99
src = ["llm_feedback_classifier"]
include = ["pyproject.toml", "llm_feedback_classifier/**/*.py"]

[tool.ruff.lint]
extend-select = ["I"]  # Add import sorting

[tool.ruff.lint.isort]
known-first-party = ["llm_feedback_classifier"]
force-sort-within-sections = true

